# horton/go/models/model_job_report.go
/*
 * Horton
 * John Shields
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package models

type JobReport struct {
	JobReportId int32 `json:"jobReportId,omitempty"`

	Date string `json:"date,omitempty"`

	VehicleModel string `json:"vehicleModel,omitempty"`

	VehicleReg string `json:"vehicleReg,omitempty"`

	VehicleLocation string `json:"vehicleLocation,omitempty"`

	MilesOnVehicle int32 `json:"milesOnVehicle,omitempty"`

	Warranty bool `json:"warranty,omitempty"`

	Breakdown bool `json:"breakdown,omitempty"`

	CustomerName string `json:"customerName,omitempty"`

	Complaint string `json:"complaint,omitempty"`

	Cause string `json:"cause,omitempty"`

	Correction string `json:"correction,omitempty"`

	Parts string `json:"parts,omitempty"`

	WorkHours int32 `json:"workHours,omitempty"`

	WorkerName string `json:"workerName,omitempty"`

	JobComplete bool `json:"jobComplete,omitempty"`

}