# repota/repotaApp/src/app/services/client_stubs/model/jobReport.ts
/**
 * Horton
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface JobReport { 
    jobReportId?: number;
    date?: string;
    vehicleModel?: string;
    vehicleReg?: string;
    milesOnVehicle?: number;
    vehicleLocation?: string;
    warranty?: string;
    breakdown?: string;
    customerName?: string;
    customerComplaint?: string;
    cause?: string;
    correction?: string;
    parts?: string;
    workHours?: number;
    workerName?: string;
    jobComplete?: string;
}